CARDS-1475: PROMs dashboard Refactored dashboard widget components to reduce code duplication Also fixed wrong link to visit subject in the Appointments box
