+++ proms-resources/frontend/src/main/frontend/src/VisitView.jsx
@@ -18,47 +18,19 @@
 //
 import React, { useState, useEffect, useContext } from "react";
 
-import questionnaireStyle from "./questionnaire/QuestionnaireStyle.jsx";
-import LiveTable from "./dataHomepage/LiveTable.jsx";
-
 import {
-  Avatar,
-  Card,
-  CardContent,
-  CardHeader,
-  Divider,
-  IconButton,
-  Tab,
-  Tabs,
-  Tooltip,
-  Typography,
   makeStyles,
-  withStyles
 } from "@material-ui/core";
 import { Link } from 'react-router-dom';
 import EventIcon from '@material-ui/icons/Event';
 
+import PromsViewInternal from "./PromsViewInternal.jsx";
+
 const useStyles = color => makeStyles(theme => ({
-  promsView : {
-    border: "2px solid " + color,
-    "& .MuiTab-root": {
-      paddingBottom: theme.spacing(1),
-      textTransform: "none",
-      fontWeight: "300",
+  visitView : {
+    "&.MuiCard-root" : {
+       border: "2px solid " + color,
     },
-    "& .MuiTab-root.Mui-selected" : {
-      color: color,
-    },
-    "& .MuiTabs-indicator": {
-      background: color,
-    },
-  },
-  promsViewAvatar: {
-    background: color,
-  },
-  promsViewTitle: {
-    fontWeight: "600",
-    fontSize: "90%",
   },
   statusIncomplete : {
     color: theme.palette.error.main,
@@ -77,39 +49,6 @@ function VisitView(props) {
 
   const classes = useStyles(color)();
 
-  const [ title, setTitle ] = useState("Appointments");
-
-  let toMidnight = (date) => {
-     date.setHours(0);
-     date.setMinutes(0);
-     date.setSeconds(0);
-     return date;
-  }
-
-  let today = new Date(), tomorrow = new Date();
-  tomorrow.setDate(today.getDate() + 1);
-  today = toMidnight(today).toISOString();
-  tomorrow = toMidnight(tomorrow).toISOString();
-
-  const tabFilter = {
-    "Past" : {
-      dateFilter :  `visitDate.value < '${today}' `,
-      order : "desc"
-    },
-    "Today" : {
-      dateFilter :  `visitDate.value >= '${today}' and visitDate.value < '${tomorrow}' `,
-      order: ""
-    },
-    "Upcoming" : {
-      dateFilter :  `visitDate.value >= '${tomorrow}' `,
-      order: ""
-    },
-  };
-
-  const tabs = Object.keys(tabFilter);
-
-  const [ activeTab, setActiveTab ] = useState(1); // Today
-
   let query = (
 "select distinct visitInformation.* " +
   "from " +
@@ -118,9 +57,9 @@ function VisitView(props) {
       "inner join [cards:Answer] as visitDate on isdescendantnode(visitDate, visitInformation) " +
   "where " +
     `visitInformation.questionnaire = '${visitInfo?.["jcr:uuid"]}' ` +
-      `and visitDate.question = '${visitInfo?.time?.["jcr:uuid"]}' and ` + tabFilter[tabs[activeTab]].dateFilter + " " +
+      `and visitDate.question = '${visitInfo?.time?.["jcr:uuid"]}' and __DATE_FILTER_PLACEHOLDER__ ` +
       `and visitSurveys.question = '${visitInfo?.surveys?.["jcr:uuid"]}' and visitSurveys.value = '${surveysId}' ` +
-  "order by visitDate.value " + tabFilter[tabs[activeTab]].order
+  "order by visitDate.value __SORT_ORDER_PLACEHOLDER__"
 )
 
   let columns = [
@@ -146,11 +85,11 @@ function VisitView(props) {
     },
     {
       "key" : "status",
-      "label" : "Status",
+      "label" : "Survey status",
       "format" : (row) => (
          <Link
            className={classes["status" + (!row.surveys_complete ? "Incomplete" : (!row.surveys_submitted ? "Unreviewed" : "Completed"))]}
-           to={`/content.html/${row.subject['@path']}`}>
+           to={`/content.html${row.subject['@path']}`}>
            { !row.surveys_complete ? "Incomplete" : (!row.surveys_submitted ? "Pending submission" : "Completed") }
          </Link>
       )
@@ -159,31 +98,16 @@ function VisitView(props) {
 
 
   return (
-    <Card className={classes.promsView}>
-      {title &&
-      <CardHeader
-        disableTypography
-        avatar={<Avatar className={classes.promsViewAvatar}><EventIcon /></Avatar>}
-        title={<Typography variant="overline" className={classes.promsViewTitle}>{title}</Typography>}
-      />
-      }
-      <Tabs value={activeTab} onChange={(event, value) => setActiveTab(value)}>
-            { tabs.map((value, index) => {
-              return <Tab label={value}  key={"form-" + index} />;
-            })}
-      </Tabs>
-      <Divider />
-      <CardContent>
-        <LiveTable
-          columns={columns}
-          customUrl={'/query?query=' + encodeURIComponent(query)}
-          defaultLimit={10}
-          questionnaire={visitInfo?.["jcr:uuid"]}
-          entryType={"Form"}
-          disableTopPagination={true}
-        />
-      </CardContent>
-    </Card>
+    <PromsViewInternal
+      className={classes.visitView}
+      color={color}
+      avatar={<EventIcon />}
+      title="Appointments"
+      columns={columns}
+      query={query}
+      dateField="visitDate"
+      questionnaireId={visitInfo?.["jcr:uuid"]}
+    />
   );
 }
